% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsol_presentation.R
\name{bsol_presentation}
\alias{bsol_presentation}
\title{Convert to an BSOL ICS xaringan presentation}
\usage{
bsol_presentation(...)
}
\arguments{
\item{...}{additional arguments to pass to [xaringan::moon_reader()]}
}
\value{
R Markdown output format to pass to \code{\link{render}}
}
\description{
Format for converting from R Markdown to a Birmingham and Solihull ICS branded xaringan
presentation.
}
\examples{
\dontrun{
library(rmarkdown)

# simple invocation
render("input.Rmd", bsol_presentation())
}
}
