% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsol_theme_cols.R
\name{bsol_theme_cols}
\alias{bsol_theme_cols}
\title{BSOL Theme Colours}
\usage{
bsol_theme_cols(
  ...,
  palette = c(NA, "main", "alternative", "ics_blue", "ics_orange", "ics_green",
    "ics_purple", "ics_navy")
)
}
\arguments{
\item{...}{individual colours that you wish to get the values of}

\item{palette}{a name of a palette to select the colours of}
}
\value{
a named vector of RGB colours in hexadecimal form
}
\description{
This function returns the colours that are used in the BSOL ICS theme as
a named vector of RGB values in hexadecimal form.
}
\details{
If no arguments are passed to the function, then all of the colours are
returned. If only specific colours are required, then the names of the
colours that are required can be passed as strings to the function, and only
those colours will be returned.

If a "pallete" is selected then no arguments should be passed to ...

Only one palete can be selected at a time.
}
\examples{
# show all of the colours
bsol_theme_cols()
# or, just show some colours
bsol_theme_cols("green", "orange", "deep_navy", "nhs_blue")
# or, select a single palette
bsol_theme_cols(palette = "slates")
}
