% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsol_theme_gradient_colour_scales.R
\name{bsol_theme_gradient_colour_scales}
\alias{bsol_theme_gradient_colour_scales}
\alias{scale_colour_bsol}
\alias{scale_color_bsol}
\alias{scale_fill_bsol}
\title{BSOL ICS Gradient colour Scale}
\usage{
scale_colour_bsol(
  palette = c("main", "alternative", "ics_blue", "ics_orange", "ics_green", "ics_purple",
    "ics_navy"),
  discrete = TRUE,
  reverse = FALSE,
  ...
)

scale_color_bsol(
  palette = c("main", "alternative", "ics_blue", "ics_orange", "ics_green", "ics_purple",
    "ics_navy"),
  discrete = TRUE,
  reverse = FALSE,
  ...
)

scale_fill_bsol(
  palette = c("main", "alternative", "ics_blue", "ics_orange", "ics_green", "ics_purple",
    "ics_navy"),
  discrete = TRUE,
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{The name of palette to use}

\item{discrete}{Optional: boolean to indicate that this scale is a discrete
scale. Defaults to TRUE}

\item{reverse}{Optional: boolean to reverse the direction of the scale.
Defaults to FALSE}

\item{...}{additional arguments passed to the ggplot functions}
}
\value{
a scale object to be used with a ggplot object
}
\description{
Generates a two colour gradient (low-high)
}
