% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsol_theme_pal.R
\name{bsol_theme_pal}
\alias{bsol_theme_pal}
\title{BSOL ICS Theme Palettes}
\usage{
bsol_theme_pal(
  palette = c("main", "alternative", "ics_blue", "ics_orange", "ics_green", "ics_purple",
    "ics_navy"),
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{The name of palette to use}

\item{reverse}{Optional: boolean value to reverse the palette}

\item{...}{arguments to pass to colorRamp}
}
\value{
A colorRamp function
}
\description{
This function returns a palette generated by colorRampPalette from one of the
BSOL palettes
}
\examples{
bsol_theme_pal()
# or, specify a palette
bsol_theme_pal("oranges")
# a reverse palette
bsol_theme_pal("slates", TRUE)
}
