% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_helpers.R
\name{ggplotly_bsol}
\alias{ggplotly_bsol}
\title{ggplotly BSOL wrapper for passing the font size properly}
\usage{
ggplotly_bsol(p = ggplot2::last_plot(), font, ...)
}
\arguments{
\item{p}{a ggplot object, default is the last plot called}

\item{font}{required font, extracted from the plot if not specified otherwise}

\item{...}{Arguments to the layout object. For documentation,
see \url{https://plotly.com/r/reference/#Layout_and_layout_style_objects}}
}
\value{
A plotly object
}
\description{
ggplotly BSOL wrapper for passing the font size properly
}
\examples{
library(ggplot)
a<-ggplot(iris, aes(y=Sepal.Length, x=Sepal.Width, colour=Species))+
geom_point()+
scale_colour_bsol()+
theme_bsol()

ggplotly_bsol(a)

}
