% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsol_theme.R
\name{theme_bsol}
\alias{theme_bsol}
\title{Creates the BSOL Theme for ggplot}
\usage{
theme_bsol(base_family = "Arial")
}
\arguments{
\item{base_family}{The font to use in this theme}
}
\description{
Creates a base theme for ggplot objects.
}
\examples{
\dontrun{
library(ggplot2)
library(BSOLTheme)
ggplot(iris, aes(Sepal.Width, Sepal.Length, colour = Species)) +
  geom_point() +
  theme_bsol()
}
}
