% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsol_theme.R
\name{theme_bsol_officer}
\alias{theme_bsol_officer}
\title{Creates the BSOL Theme for ggplot, with Officer package}
\usage{
theme_bsol_officer(base_family = "Open Sans")
}
\arguments{
\item{base_family}{The font to use in this theme}
}
\description{
Creates a base theme for ggplot objects, but some of the sizes needed boosting to work with exporting
to Powerpoint using Officer.  When translating over, the size gets a bit lost.  This works with the
template included in this package.
}
\examples{
\dontrun{
library(ggplot2)
library(BSOLTheme)
ggplot(iris, aes(Sepal.Width, Sepal.Length, colour = Species)) +
  geom_point() +
  theme_bsol_officer()
}
}
